:recycle: Refactor: #33 - SearchResult 모델 분리
SearchResult 파일에 있던 모델들을 별도의 파일로 분리합니다.
