+++ Bookie/Bookie/Models/SearchResult.swift
@@ -29,37 +29,6 @@ struct SearchResult: Codable {
   }
 }
 
-// MARK: - Book
-
-struct Book: Codable {
-  enum CodingKeys: String, CodingKey {
-    case title, link, author, pubDate, description, isbn, isbn13
-    case itemID = "itemId"
-    case priceSales, priceStandard, mallType, stockStatus, mileage
-    case coverURL = "cover"
-    case categoryID = "categoryId"
-    case categoryName, publisher, salesPoint, adult, customerReviewRank, seriesInfo, subInfo, fixedPrice
-  }
-
-  let title: String
-  let link: String
-  let author, pubDate, description, isbn: String
-  let isbn13: String
-  let itemID, priceSales, priceStandard: Int
-  let mallType: MallType
-  let stockStatus: String
-  let mileage: Int
-  let coverURL: String
-  let categoryID: Int
-  let categoryName, publisher: String
-  let salesPoint: Int
-  let adult: Bool
-  let customerReviewRank: Int
-  let seriesInfo: SeriesInfo?
-  let subInfo: SubInfo?
-  let fixedPrice: Bool?
-
-}
 
 // MARK: - MallType
 
@@ -70,24 +39,4 @@ enum MallType: String, Codable {
   case dvd = "DVD"
   case used = "USED"
   case ebook = "EBOOK"
-}
-
-// MARK: - SeriesInfo
-
-struct SeriesInfo: Codable {
-  let seriesID: Int
-  let seriesLink: String
-  let seriesName: String
-
-  enum CodingKeys: String, CodingKey {
-    case seriesID = "seriesId"
-    case seriesLink, seriesName
-  }
-}
-
-// MARK: - SubInfo
-
-struct SubInfo: Codable {
-  let subTitle, originalTitle: String?
-  let itemPage: Int?
 }
