+++ Bookie/Bookie/Models/SeriesInfo.swift
@@ -5,4 +5,16 @@
 //  Created by 박제균 on 2/8/24.
 //
 
-import Foundation
+// MARK: - SeriesInfo
+
+struct SeriesInfo: Codable {
+  let seriesID: Int
+  let seriesLink: String
+  let seriesName: String
+
+  enum CodingKeys: String, CodingKey {
+    case seriesID = "seriesId"
+    case seriesLink, seriesName
+  }
+}
+
