+++ Bookie/Bookie/Models/SubInfo.swift
@@ -5,4 +5,8 @@
 //  Created by 박제균 on 2/8/24.
 //
 
-import Foundation
+// MARK: - SubInfo
+struct SubInfo: Codable {
+  let subTitle, originalTitle: String?
+  let itemPage: Int?
+}
